HWMonitorSMC2.app v2.6.0
Promotion to a greater minor version. Latest commits introduces a huge change for Intel Power Gadget, which introduce a new header (so new code), not last the ability to show frequencies and temperatures for a multi CPU-Package Systems.
Implemented support for SMCSuperIO.kext which doesn't publish smc keys. A little database was created for ASUS and Gigabyte motherboards configurations easily expandable. Configurations are autodetected.
The UI has now the ability to detach singles sensors. Charts (or plots) expanded for voltages, amperages etc..
Minor fixes and improvements.